+++ src/main/java/com/lama/dal/service/OrderService.java
@@ -1,6 +1,7 @@
 package com.lama.dal.service;
 
 import com.lama.dal.entity.Order;
+import com.lama.dal.error.EntityNotFoundException;
 import com.lama.dal.repository.OrderRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,14 +21,30 @@ public class OrderService {
     }
 
     public Optional<Order> findById(String id) {
-        return orderRepository.findById(id);
+        Optional<Order> order = orderRepository.findById(id);
+        if (!order.isPresent()) {
+            throw new EntityNotFoundException(Order.class, "id", id);
+        }
+
+        return order;
+    }
+
+    public Order update(String id, Order order) {
+        if (!findById(id).isPresent()) {
+            throw new EntityNotFoundException(Order.class, "id", id);
+        }
+        return orderRepository.save(order);
     }
 
     public Order save(Order order) {
-       return (Order) orderRepository.save(order);
+        return orderRepository.save(order);
     }
 
     public void deleteById(String id) {
+        if (!findById(id).isPresent()) {
+            throw new EntityNotFoundException(Order.class, "id", id);
+        }
+
         orderRepository.deleteById(id);
     }
 }
